requirements.txt
annotated-types==0.7.0
anyio==4.4.0
bcrypt==4.2.0
beanie==1.26.0
cffi==1.17.1
click==8.1.7
cryptography==43.0.1
dnspython==2.6.1
ecdsa==0.19.0
email_validator==2.2.0
fastapi==0.114.1
h11==0.14.0
idna==3.8
itsdangerous==2.2.0
Jinja2==3.1.4
jose==1.0.0
jwt==1.3.1
lazy-model==0.2.0
MarkupSafe==2.1.5
motor==3.5.1
passlib==1.7.4
pyasn1==0.6.1
pycparser==2.22
pydantic==2.9.1
pydantic-settings==2.5.2
pydantic_core==2.23.3
pymongo==4.8.0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
rsa==4.9
six==1.16.0
sniffio==1.3.1
starlette==0.38.5
toml==0.10.2
typing_extensions==4.12.2
uvicorn==0.30.6
